refactor: use typescript for gatsby-node, gatsby-ssr and gatsby-browser
