+++ gatsby-ssr.tsx
@@ -3,9 +3,9 @@ import React from 'react'
 import Layout from './src/components/Layout'
 import WrapWithProvider from './src/hoc/WrapWithProvider'
 
-export const wrapPageElement = ({ element }) => {
-  // props provide same data to Layout as Page element will get
-  // including location, data, etc - you don't need to pass it
+type WrapPageElementsProps = ({ element }: { element: React.ReactNode }) => React.ReactNode
+
+export const wrapPageElement: WrapPageElementsProps = ({ element }) => {
   return <Layout>{element}</Layout>
 }
